adds func errors, use gogoproto
this adds the concept of func errors, which are errors that can be blamed on a
function rather than the service, eg for things like a function hitting the
timeout due to being inside the container for too long to execute. to do this,
need to bubble up this field through grpc as these errors need to be
distinguished runner-side but consumed lb-side for metrics. this adds a new
type that builds on api errors, with the added distinction for function
specific errors, and transmits over grpc to rebuild the error's type on the lb
side.

this also updates the protobufs to use gogo proto. see gogo proto repo for why
this is a good idea, but we get a lot of benefits, grpc will be a lot faster
with lower memory consumption namely (aside from some aesthetics for more
canonical go structs, which we care less about since we pretty much have
isolated usage of our grpc structs). the reason here honestly is that i
couldn't get the normal protobufs working and i spent less time getting the
gogo protobufs working than the normal ones, where tests would fail. will need
to think about roll out strategy here and/or test, these structs should be
compatible one way or the other but new lbs may need to start with 0 old
runners and add only new runners - TBD here, will test. sorry that this bloats
the patch quite a bit, see motivation above, but we should move to these imo
as we're doing only go to go protobufs in our use case. there are other things
like combing through fields we can mark as nullable to save lots of allocs on
pointers (most of our fields on our structs appear to always be set, this is
easy enough. TODO here)

closes #1359